+++ src/main/java/com/wanted/preonboarding/ticketing/controller/PerformanceController.java
@@ -1,0 +1,25 @@
+package com.wanted.preonboarding.ticketing.controller;
+
+import com.wanted.preonboarding.ticketing.domain.dto.response.ReadPerformanceResponse;
+import com.wanted.preonboarding.ticketing.service.PerformanceService;
+import lombok.RequiredArgsConstructor;
+import org.springframework.data.domain.Page;
+import org.springframework.data.domain.Pageable;
+import org.springframework.data.domain.Sort;
+import org.springframework.data.web.PageableDefault;
+import org.springframework.http.ResponseEntity;
+import org.springframework.web.bind.annotation.GetMapping;
+import org.springframework.web.bind.annotation.RequestParam;
+import org.springframework.web.bind.annotation.RestController;
+
+@RestController
+@RequiredArgsConstructor
+public class PerformanceController {
+    private final PerformanceService performanceService;
+
+    @GetMapping("/performance")
+    public ResponseEntity<Page<ReadPerformanceResponse>> readPerformance(@RequestParam String isReserve,
+                                                                         @PageableDefault(sort = "id", direction = Sort.Direction.DESC) Pageable pageable) {
+        return ResponseEntity.ok(performanceService.readPerformance(isReserve, pageable));
+    }
+}
