+++ src/main/java/com/wanted/preonboarding/ticketing/domain/dto/PerformanceType.java
@@ -1,0 +1,17 @@
+package com.wanted.preonboarding.ticketing.domain.dto;
+
+import lombok.Getter;
+
+@Getter
+public enum PerformanceType {
+    NONE(0),
+    CONCERT(1),
+    EXHIBITION(2);
+
+    private final int category;
+
+    PerformanceType(int category) {
+        this.category = category;
+    }
+
+}
